## Multiple commits
There are two commits here. They are both important.

Commit 1: Refactors the existing code so that each relation type is fetched in its own method. There's more information about why I needed this in the commit message.
Commit 2: This is the commit that holds the actual change of logic. This commit also includes changes to some variable names, to make them more explicit or self explanatory.

It'll almost certainly be easier to review this PR by looking at the changes _in each commit_, so you can more easily see the actual change to logic. But for convenience's sake, I've added a github comment to every change in the second commit that _isn't_ just a change of variable name or linting fix or added comment.

## Performance before and after
The following results are for effectively the same setup as in the linked issue - except the `Team` class also has a separate `has_many` relation for the `Player` class, which holds another (separate) 10 players per team. This means there are a total of 20,000 player records in the database.

The following numbers are for the full request of loading a page, and come from the lekoala debugbar, so they're not 100% accurate - but they give a clear indication of the improvement this PR makes.

### has_many
This PR doesn't affect the performance of eager loading has_many relations - but it's nice to see that eagerloading does improve the speed of loading has_many relations in general.
|without eager loading|with eager loading (no PR)|with eager loading (with PR)|
|---|---|---|
|~400ms|~350ms|~350ms|

### many_many
Without this PR, eager loading many_many relations is effectively impossible.
Even with this PR, we can see that eager loading many_many relations is _slower_ than without it, so there's still room for further optimisations (it's worth noting the database query itself (`WHERE ("Player"."ID" IN (<IDs of teams>)`) is taking ~1 second of that time so the PHP side of it is about as good as it's likely to get).
|without eager loading|with eager loading (no PR)|with eager loading (with PR)|
|---|---|---|
|~500ms|times out (30+ seconds)|~1.5seconds|

## Issue
- https://github.com/silverstripe/silverstripe-framework/issues/10847