@emteknetnz 2023-07-06
Just noting in my own testing with this PR that eager loading actually had a larger impact on many-many than has-many, counter to the testing in the description on this PR.  I guess it's a case of "your mileage may vary" and "you should validate whether or not eager loading works for your particular website"